core/contents/requirements.txt
Django
Pillow
django-ckeditor
django-content-editor
django-versatileimagefield
--upgrade pip
